+++ CHANGES.md
@@ -7,4 +7,4 @@ V 2.2.0
 * Python3 compat fixes
 * remove update
 * remove dreambox specific functions
-
+* rename myplex to plex.tv